Bugfix and parallelization for LowCID precursor deconvolution:

Code | Data | "GetMs2Scans" Runtime | Max RAM Usage | PSMs
-- | -- | -- | -- | --
master | LowResHela | 93.0 sec | 7.8 GB | 7959
PR | LowResHela | 10.4 sec | 1.2 GB | 7959
  |   |   |   |  
master | HighResMouse | 10.9 sec | 1.9 GB | 14969
PR | HighResMouse | 11.0 sec | 1.9 GB | 14969